- Dockerfile and docker-compose.yml - Moved source files into a src folder - Removed hashed_password from UserService.list
