+++ docker-compose.yaml
@@ -1,0 +1,33 @@
+version: '3.8'
+
+services:
+  mongo:
+    image: mongo
+    restart: always
+    environment:
+      MONGO_INITDB_ROOT_USERNAME: ${MONGODB_USERNAME}
+      MONGO_INITDB_ROOT_PASSWORD: ${MONGODB_PASSWORD}
+    ports:
+      - 27017:27017
+    volumes:
+      - ./data:/data/db
+    networks:
+      - backend
+
+  backend:
+    build: .
+    image: hue-backend
+    environment:
+      MONGODB_HOSTNAME: mongo
+      MONGODB_PORT: ${MONGODB_PORT}
+    ports:
+      - 8080:80
+    depends_on:
+      - mongo
+    networks:
+      - backend
+
+volumes:
+  data:
+networks:
+  backend:
